+++ src/components/Search/Search.tsx
@@ -17,7 +17,6 @@ import {
   Form,
   useModal,
   TemplateFooter,
-  Footer,
   useNYPLBreakpoints,
 } from "@nypl/design-system-react-components";
 import { useRouter } from "next/router";
@@ -341,9 +340,7 @@ const SearchResults: React.FC<{
           />
         </TemplateContentPrimary>
       </TemplateContent>
-      <TemplateFooter>
-        <Footer />
-      </TemplateFooter>
+      <TemplateFooter />
     </Template>
   );
 };