+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOG
 
+## [0.17.1]
+- Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app 
+- Refactored Playwright/Cucumber tests to use new click action function
+
 ## [0.17.0]
 - Replace Search Examples with Recently Added Collections to landing page
 - Add primary hero to the landing page
@@ -7,8 +11,6 @@
 - Add Collection page
 - Add collection placeholder image
 - Update hero image to update based on screen size
-- Chore: fresh package-lock.json install and fix install errors on mac with M1 chip
-- Add Playwright/Cucumber tests to support new Recently Added Collections feature on landing page
 - Chore: fresh package-lock.json install and fix install errors on mac with M1 chip
 - Add Playwright/Cucumber tests to support new Recently Added Collections feature on landing page
 - Refactored Playwright/Cucumber tests to use new page visit actions function